fix(grpc): report persisted work package counts (#1069)
* fix(grpc): report persisted work package counts

* fix(grpc): expose persisted work package dependencies

* ci: isolate cargo audit home

* fix(ci): prepare musl fuzz sanitizer target

* fix(proto): preserve large work package dependency ids

---------

+++ crates/agileplus-proto/src/stubs.rs
@@ -96,7 +96,7 @@ pub struct WorkPackageStatus {
     pub agent_id: String,
     pub pr_url: String,
     pub pr_state: String,
-    pub depends_on: Vec<i32>,
+    pub depends_on: Vec<i64>,
     pub file_scope: Vec<String>,
 }
